added devices and projecttype components :boom:
