+++ src/projects/reducers/project.js
@@ -67,7 +67,7 @@ export const newProject = modelReducer('newProject', {
   utm: {
     code: ''
   },
-  type: '',
+  type: 'visual_design',
   details: {
     version: 'v2',
     devices: ['phone'],
